v9.10.77: MRSA Phasen 06/20/24/29/56, PGHI-Fixes, MOS-Gate mp3/aac 4.5, utils.py Bereinigung, Griffin-Lim-Elimination
- MRSA 5-Zonen-Implementierung in phase_06, phase_20, phase_24, phase_29, phase_56
- P1-Fix: audiosr_plugin._spectral_exciter deterministisch + PGHI statt Griffin-Lim (Sec 2.40 + 4.5)
- P1-Fix: aurik_denker mp3_high/aac MOS-Ziel 4.2 -> 4.5 (Sec 6.2)
- P2-Fix: masking_remover iSTFT+Warning statt stiller Griffin-Lim (Sec 4.5)
- P2-Fix: vocos_plugin iSTFT mit Originalphase statt Griffin-Lim + random phase (Sec 2.40 + 4.5)
- utils.py: 13 Fehler behoben (logging, determinismus, print->logger, ipdb entfernt, cudnn.benchmark=False)
- .gitignore: Laufzeit-Artefakte ergaenzt (output/, reports/, pip_*.txt, *_results.txt)
- Neue normative Tests: test_competitive_stratified_gate, test_external_mushra_artifact_contract,
  test_full_pipeline_determinism, test_stability_invariants
- Neue Unit-Tests: test_adaptive_stft_consistency, test_auto_eq, test_dynamic_range_expander,
  test_harmonic_exciter, test_multiband_dynamics_modules, test_soundstage_depth_enhancer
- Neue Module: phase_output_guard.py, aliasing.py, run_aurik.sh, pytest_summary_guard.sh
- ruff auto-format: 70 Dateien reformatiert (pre-commit hook)

+++ backend/core/phases/phase_38_presence_boost.py
@@ -304,8 +304,8 @@ class PresenceBoost(PhaseInterface):
         a2 = 1 - alpha / A
 
         # Normalize
-        b = np.array([b0, b1, b2]) / a0
-        a = np.array([1, a1 / a0, a2 / a0])
+        b = np.array([b0, b1, b2], dtype=np.float64) / a0
+        a = np.array([1.0, a1 / a0, a2 / a0], dtype=np.float64)
 
         # Apply filter
         filtered = signal.lfilter(b, a, audio)
